+++ hspec-core/test/Test/Hspec/Core/Formatters/DiffSpec.hs
@@ -1,15 +1,40 @@
+{-# LANGUAGE ScopedTypeVariables #-}
 module Test.Hspec.Core.Formatters.DiffSpec (spec) where
+
+import           Prelude ()
+import           Test.Hspec.Core.Compat
 
 import           Helper
 import           Data.Char
 
 import           Test.Hspec.Core.Formatters.Diff
 
+dropQuotes :: String -> String
+dropQuotes = init . tail
+
 spec :: Spec
 spec = do
   describe "partition" $ do
-    it "puts backslash-escaped characters into a separate chunks" $ do
-      partition (show "foo\nbar") `shouldBe` ["\"", "foo", "\\n", "bar", "\""]
+    context "with a single shown Char" $ do
+      it "never partitions a character escape" $ do
+        property $ \ (c :: Char) -> partition (show c) `shouldBe` ["'", dropQuotes (show c), "'"]
+
+    context "with a shown String" $ do
+      it "puts backslash-escaped characters into separate chunks" $ do
+        partition (show "foo\nbar") `shouldBe` ["\"", "foo", "\\n", "bar", "\""]
+
+      it "puts *arbitrary* backslash-escaped characters into separate chunks" $ do
+        property $ \ xs c ys ->
+          let
+            char = dropQuotes (show [c])
+            isEscaped = length char > 1
+            escape = tail char
+            sep = case ys of
+              x : _ | all isDigit escape && isDigit x || escape == "SO" && x == 'H' -> ["\\&"]
+              _ -> []
+            actual = partition (show (xs ++ c : ys))
+            expected = partition (init $ show xs) ++ [char] ++ sep ++ partition (tail $ show ys)
+          in isEscaped ==> actual `shouldBe` expected
 
   describe "breakList" $ do
     context "with a list where the predicate matches at the beginning and the end" $ do